This is just a test to get familiar with git and a few questions.
train_instructions.py seems to be equivalent to original code train_hierarchy but only implementing SimpleNetwork (which is the "instruction_model import InstructionModel" here).
So, I wonder do we miss the following two lines in class StateEncoderA in instruction_model.py?
    self.fc6 = nn.Linear(48+32, 48)
    self.fc7 = nn.Linear(48, 9)   #it was 8
-- figured out. These two lines were included in class InstructionModel. But it is 48 rather than 48+32 ?

Also, the results seem to be different from the original code.

One other question: the "instructions" in line 70 is not used; should this be included in ir_generator (InstructionModel)?